fixed unit test failures
