+++ go.mod
@@ -4,5 +4,7 @@ go 1.14
 
 require (
 	github.com/gorilla/mux v1.8.0
-	github.com/spf13/viper v1.7.1 // indirect
+	github.com/julienschmidt/httprouter v1.2.0
+	github.com/spf13/viper v1.7.1
+	github.com/stretchr/testify v1.3.0
 )
